move HomePageImage to its own component
